+++ backend/src/main/java/is/hi/moviedb/service/UserService.java
@@ -16,6 +16,6 @@ public interface UserService {
     
     List<User> getAllUsers();
 
-    Long getUserId(String email);
+    int getUserId(String email);
 
 }
