+++ backend/src/main/java/is/hi/moviedb/service/implementations/ReviewServiceImpl.java
@@ -36,12 +36,12 @@ public class ReviewServiceImpl implements ReviewService{
      * Creates a new Review
      * @param userId
      * @param mediaId
-     * @param movieReview
+     * @param mediaReview
      * @param rating
      * @return Review
      */
     @Override
-    public Review createReview(long userId, long mediaId, String mediaReview, double rating){
+    public Review createReview(int userId, int mediaId, String mediaReview, double rating){
         Review review = new Review(userId,mediaId,mediaReview,rating);
         reviewRepository.save(review);
         return review;
@@ -56,7 +56,7 @@ public class ReviewServiceImpl implements ReviewService{
      * @return Review
      */
     @Override
-    public Review createSeasonReview(long userId,long seasonId,String seasonReview,double rating){
+    public Review createSeasonReview(int userId,int seasonId,String seasonReview,double rating){
 
         //Create the review for the season
         Review review = new Review(userId,seasonId,seasonReview,rating);
@@ -66,7 +66,7 @@ public class ReviewServiceImpl implements ReviewService{
         List<Episode> episodeList = episodeService.getEpisodesBySeasonId((int) seasonId);
         int n = episodeList.size();
         for (int i = 0; i < n;i++) {
-            long episodeId = episodeList.get(i).getId();
+            int episodeId = episodeList.get(i).getId();
             if (findByUserIdAndMediaId(userId,episodeId) == null) {
                 createReview(userId,episodeId,"",rating);
             }
@@ -83,7 +83,7 @@ public class ReviewServiceImpl implements ReviewService{
      * @return Review
      */
     @Override
-    public Review createTvShowReview(long userId,long tvShowId,String tvShowReview,double rating){
+    public Review createTvShowReview(int userId,int tvShowId,String tvShowReview,double rating){
 
         //Create the review for the TvShow
         Review review = new Review(userId,tvShowId,tvShowReview,rating);
@@ -93,7 +93,7 @@ public class ReviewServiceImpl implements ReviewService{
         List<Season> seasonList = seasonService.getSeasonsByTvShowId((int) tvShowId);
         int n = seasonList.size();
         for (int i = 0; i < n; i++) {
-            long seasonId = seasonList.get(i).getId();
+            int seasonId = seasonList.get(i).getId();
             if (findByUserIdAndMediaId(userId,seasonId) == null) {
                 createSeasonReview(userId,seasonId,"",rating);
             }
@@ -126,7 +126,7 @@ public class ReviewServiceImpl implements ReviewService{
      */
 
     @Override
-    public Review changeReview(long userId,long mediaId,String mediaReview,double rating){
+    public Review changeReview(int userId,int mediaId,String mediaReview,double rating){
         Review review = reviewRepository.findById(Long.toString(userId) + "-" + Long.toString(mediaId));
         review.setRating(rating);
         review.setMediaReview(mediaReview);
@@ -144,7 +144,7 @@ public class ReviewServiceImpl implements ReviewService{
      * @return Review
      */
     @Override
-    public Review findByUserIdAndMediaId(long userId,long mediaId){
+    public Review findByUserIdAndMediaId(int userId,int mediaId){
         Review review = reviewRepository.findByUserIdAndMediaId(userId,mediaId);
         if (review != null){
             return review;
@@ -172,7 +172,7 @@ public class ReviewServiceImpl implements ReviewService{
      * @return List of Reviews
      */
     @Override
-    public List<Review> findByUserId(long userId){
+    public List<Review> findByUserId(int userId){
         List<Review> reviews = reviewRepository.findByUserId(userId);
         if (reviews != null){
             return reviews;
@@ -186,7 +186,7 @@ public class ReviewServiceImpl implements ReviewService{
      * @return List of Reviews
      */
     @Override
-    public List<Review> findByMediaId(long mediaId){
+    public List<Review> findByMediaId(int mediaId){
         List<Review> reviews = reviewRepository.findByMediaId(mediaId);
         if (reviews != null){
             return reviews;
